Implement can_connect DBUS method
can_connect simply attempts to open a connection to a TACACS+ server
to determine whether a server is reachable.

Since the same logic is used as when we connect to perform a TACACS+
transaction, can_connect respects any active hold down timers, and
can also cause hold down timers to be activated and tacplusd to go
into offline mode.

VRVDR-50035