tacplusd gets stuck in offline mode
While iterating the server list, in tacplus_connect(), it is possible
for hold down timers to expire. This is particularly likely with
large server lists or when high timeout values are in use.

After iterating the server list and determining that all servers are
held down we did not check whether the next expiring hold down timer
had already expired. The result of this was that tacplusd was placed
into offline mode but no timer was set to bring it out again, so
tacplusd remained in the offline state.

VRVDR-51238