Always attempt to get remote IP for transaction
The ut_host member of the utmpx struct may contain a hostname, but we
wish to always send an IP address in the rem_addr field of TACACS+
requests.

Therefore we now check whether the value in ut_host is an IP address,
by checking the success or failure of calls to inet_pton() passing
AF_INET and AF_INET6. If both calls fail we assume that ut_host is
either unset or is a hostname. In this case we instead obtain the IP
address from ut_addr_v6, using inet_ntop(). If this fails to yield
an IP address the best we can do is return ut_host as before, if it
was set.

Since there is no indication of the address family of the IP address
stored in ut_addr_v6, we assume the address is IPv4 if elements 1, 2,
and 3 are 0, otherwise we assume IPv6. This should work in most cases
but we cannot assert that the address will be formatted correctly in
all cases.

We could configure sshd to never store a hostname in ut_host which
should address the issue of needing to fallback to use ut_addr_v6.
This can be done by passing the "-u0" argument to sshd - see sshd(8).

However since this is a behaviour change which affects other outputs
and logs in the system, not just TACACS+, this approach is not taken
at the present time. The changes in this commit do not, however,
preclude it in the future.

VRVDR-52120