Don't exit(1) if no servers are configured
Previously, loading a configuration file with no servers defined
would cause tacplusd to exit. Although this is an unusual scenario
it is not really an error condition and exiting non-zero, triggering
systemd to restart us, is an over reaction.

In future it will be possible to hit this condition when the ability
to disable TACACS+ servers is added to the DANOS configuration model,
if all configured servers are disabled.

VRVDR-42094