Implement get_account_task_id() DBus method
If a task_id argument is not provided to an account_send() or
cmd_account_send() DBus method call we allocate a unique ID from
a global incrementing counter.

In order to allow multiple calls to the account methods for a single
task, eg. to support start/stop records, we must provide a capability
to allocate and retrieve such an ID so that the caller can pass the
same ID to all relevant method calls.

This has been implemented via a new method as an optimisation. By
decoupling the ID allocation from the account methods, ie. doing an
accounting transaction, the ID allocation can be handled entirely by
the bus processing thread so it is not delayed by any ongoing
transactions.

Clients wishing to use this functionality can therefore allocate and
cache a task ID prior to performing the actual accounting calls.

VRVDR-51580