Store hold down interval value on activation
When activating the hold down timer for a given server, take a copy
of the currently configured hold down timer interval and store it in
the server state structure. This copied value is then used when
determining if a hold down timer is active.

This change allows more flexible behaviour when reloading the
configuration. If the hold down interval is changed for a server,
and the timer is running for that server, we can choose whether
to apply the value immediately (by immediately copying the new
configuration to the state structure) or wait for the next hold
down timer activation.