parser.c: NULL GError pointer after use
GLib logs a warning if library functions are passed a non-null GError
pointer:

  GError set over the top of a previous GError or uninitialized memory.
  This indicates a bug in someone's code. You must ensure an error is NULL before it's set.
  The overwriting error message was: Key file does not have key “SourceAddress” in group “server1”

Therefore ensure that the pointer is set to NULL after each error has
been handled. This is mostly done by updating g_syslog() to take care
of it, however explicit calls are still required in a couple of
places.