Record timestamp of hold-down timer resets
When resetting a server's hold-down timer, rather than clearing the
time at which the timer was activated, store the current time in a
per-server state variable. When determining whether the timer is
active for a server, in tacplus_server_remaining_hold_down_at(),
we then check whether the server last experienced a failure before
the time of last reset; if so the timer is deemed not active.

This re-work permits resetting the hold-down timer of a server from
another thread, as it removes the possibility of multiple writers to
lastTrouble.