utils: ut_host may not be NUL terminated
If the value returned in ut_host, from getutxline(), is of length
UT_HOSTSIZE it will not be NUL terminated; get_tty_login_addr()
currently assumes it is.

From utmp(5):
> String fields are terminated by a null byte ('\0') if
> they are shorter than the size of the field.

Update get_tty_login_addr() to ensure we handle a NUL terminated
string, and add a UT to verify the fix.

VRVDR-52997